desktop: Rename config.rs to preferences.rs
