+++ desktop/src/main.rs
@@ -9,11 +9,11 @@
 mod app;
 mod backends;
 mod cli;
-mod config;
 mod custom_event;
 mod executor;
 mod gui;
 mod player;
+mod preferences;
 mod task;
 mod time_demo;
 mod util;
